refactor the strategic merge patch transformer toward moving it to a plugin
