+++ k8sdeps/transformer/factory.go
@@ -6,6 +6,7 @@ package transformer
 
 import (
 	"sigs.k8s.io/kustomize/v3/k8sdeps/transformer/patch"
+	"sigs.k8s.io/kustomize/v3/pkg/resmap"
 	"sigs.k8s.io/kustomize/v3/pkg/resource"
 	"sigs.k8s.io/kustomize/v3/pkg/transformers"
 )
@@ -16,6 +17,12 @@ type FactoryImpl struct{}
 // NewFactoryImpl makes a new factoryImpl instance
 func NewFactoryImpl() *FactoryImpl {
 	return &FactoryImpl{}
+}
+
+func (p *FactoryImpl) MergePatches(patches []*resource.Resource,
+	rf *resource.Factory) (
+	resmap.ResMap, error) {
+	return patch.MergePatches(patches, rf)
 }
 
 // MakePatchTransformer makes a new patch transformer
